+++ src/index.ts
@@ -3,17 +3,19 @@ import { FetchError } from 'named-app-errors'
 
 import type { SerializedValue } from '@ergodark/types'
 
+export type FetchConfig = Omit<RequestInit, 'body'> & {
+    rejects?: boolean,
+    ignoreParseErrors?: boolean,
+    body?: Record<string, unknown>
+};
+
 /**
  * The default `config` all fetch() calls use by default. Will be merged
  * (overridden) with the `config` object passed into each call to fetch(), if
  * provided. See [unfetch](https://github.com/developit/unfetch) for valid
  * config keys.
  */
-let globalFetchConfig: Omit<RequestInit, 'body'> & {
-    rejects?: boolean,
-    ignoreParseErrors?: boolean,
-    body?: Record<string, unknown>
-} = {
+let globalFetchConfig: FetchConfig = {
     method: 'POST',
     // credentials: 'include', // ? If you want to send and receive cookies
     headers: { 'Content-Type': 'application/json' },
@@ -35,7 +37,7 @@ export function getGlobalFetchConfig() {
 /**
  * Set the default config object merged in during all fetch() calls.
  */
-export function setGlobalFetchConfig(config: typeof globalFetchConfig) {
+export function setGlobalFetchConfig(config: FetchConfig) {
     globalFetchConfig = config;
 }
 
@@ -46,7 +48,7 @@ export function setGlobalFetchConfig(config: typeof globalFetchConfig) {
  *
  * Returns an HTTP Response object `res` and parsed response body `json`.
  */
-export async function fetch<JsonType extends SerializedValue>(url: RequestInfo, config?: typeof globalFetchConfig) {
+export async function fetch<JsonType extends SerializedValue>(url: RequestInfo, config?: FetchConfig) {
     const parsedOptions: RequestInit = {
         ...getGlobalFetchConfig(),
         ...config,
@@ -68,21 +70,21 @@ export async function fetch<JsonType extends SerializedValue>(url: RequestInfo, 
 /**
  * Syntactic sugar for calling `fetch(..., { method: 'GET', ... })`.
  */
-fetch.get = <JsonType extends SerializedValue>(url: RequestInfo, config?: typeof globalFetchConfig) => {
+fetch.get = <JsonType extends SerializedValue>(url: RequestInfo, config?: FetchConfig) => {
     return fetch<JsonType>(url, { method: 'GET', ...config });
 };
 
 /**
  * Syntactic sugar for calling `fetch(..., { method: 'PUT', ... })`.
  */
-fetch.put = <JsonType extends SerializedValue>(url: RequestInfo, config?: typeof globalFetchConfig) => {
+fetch.put = <JsonType extends SerializedValue>(url: RequestInfo, config?: FetchConfig) => {
     return fetch<JsonType>(url, { method: 'PUT', ...config });
 };
 
 /**
  * Syntactic sugar for calling `fetch(..., { method: 'DELETE', ... })`.
  */
-fetch.delete = <JsonType extends SerializedValue>(url: RequestInfo, config?: typeof globalFetchConfig) => {
+fetch.delete = <JsonType extends SerializedValue>(url: RequestInfo, config?: FetchConfig) => {
     return fetch<JsonType>(url, { method: 'DELETE', ...config });
 };