+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Changed
 - Better README.md
 - Improvements to the build process
+- Export `FetchConfig` type explicitly
 
 #### Fixed
 - Fixed CHANGELOG.md